+++ src/channels/pages/ChannelsListPage/ChannelsListPage.stories.tsx
@@ -1,3 +1,4 @@
+import { limits, limitsReached } from "@saleor/fixtures";
 import Decorator from "@saleor/storybook/Decorator";
 import { storiesOf } from "@storybook/react";
 import React from "react";
@@ -7,6 +8,7 @@ import ChannelsListPage, { ChannelsListPageProps } from "./ChannelsListPage";
 
 const props: ChannelsListPageProps = {
   channelsList,
+  limits,
   navigateToChannelCreate: () => undefined,
   onBack: () => undefined,
   onRemove: () => undefined,
@@ -16,4 +18,8 @@ const props: ChannelsListPageProps = {
 storiesOf("Views / Channels / Channels list", module)
   .addDecorator(Decorator)
   .add("default", () => <ChannelsListPage {...props} />)
-  .add("empty", () => <ChannelsListPage {...props} channelsList={[]} />);
+  .add("empty", () => <ChannelsListPage {...props} channelsList={[]} />)
+  .add("no limits", () => <ChannelsListPage {...props} limits={undefined} />)
+  .add("limits reached", () => (
+    <ChannelsListPage {...props} limits={limitsReached} />
+  ));