+++ src/products/components/ProductUpdatePage/ProductUpdatePage.tsx
@@ -16,6 +16,7 @@ import Metadata from "@saleor/components/Metadata/Metadata";
 import PageHeader from "@saleor/components/PageHeader";
 import SaveButtonBar from "@saleor/components/SaveButtonBar";
 import SeoForm from "@saleor/components/SeoForm";
+import { RefreshLimits_shop_limits } from "@saleor/components/Shop/types/RefreshLimits";
 import { ProductChannelListingErrorFragment } from "@saleor/fragments/types/ProductChannelListingErrorFragment";
 import { ProductErrorWithAttributesFragment } from "@saleor/fragments/types/ProductErrorWithAttributesFragment";
 import { TaxTypeFragment } from "@saleor/fragments/types/TaxTypeFragment";
@@ -77,6 +78,7 @@ export interface ProductUpdatePageProps extends ListActions, ChannelProps {
   fetchMoreCategories: FetchMoreProps;
   fetchMoreCollections: FetchMoreProps;
   isMediaUrlModalVisible?: boolean;
+  limits: RefreshLimits_shop_limits;
   variants: ProductDetails_product_variants[];
   media: ProductDetails_product_media[];
   hasChannelChanged: boolean;
@@ -140,6 +142,7 @@ export const ProductUpdatePage: React.FC<ProductUpdatePageProps> = ({
   media,
   hasChannelChanged,
   header,
+  limits,
   placeholderImage,
   product,
   saveButtonBarState,
@@ -320,6 +323,7 @@ export const ProductUpdatePage: React.FC<ProductUpdatePageProps> = ({
                 {hasVariants ? (
                   <ProductVariants
                     disabled={disabled}
+                    limits={limits}
                     variants={variants}
                     product={product}
                     onRowClick={onVariantShow}