+++ src/products/views/ProductUpdate/ProductUpdate.tsx
@@ -15,6 +15,7 @@ import useAppChannel from "@saleor/components/AppLayout/AppChannelContext";
 import { AttributeInput } from "@saleor/components/Attributes";
 import ChannelsAvailabilityDialog from "@saleor/components/ChannelsAvailabilityDialog";
 import NotFoundPage from "@saleor/components/NotFoundPage";
+import { useShopLimitsQuery } from "@saleor/components/Shop/query";
 import { WindowTitle } from "@saleor/components/WindowTitle";
 import { DEFAULT_INITIAL_SEARCH_DATA } from "@saleor/config";
 import { useFileUploadMutation } from "@saleor/files/mutations";
@@ -162,6 +163,11 @@ export const ProductUpdate: React.FC<ProductUpdateProps> = ({ id, params }) => {
     variables: { id }
   });
   const { channel } = useAppChannel();
+  const limitOpts = useShopLimitsQuery({
+    variables: {
+      productVariants: true
+    }
+  });
 
   const [uploadFile, uploadFileOpts] = useFileUploadMutation({});
 
@@ -234,6 +240,7 @@ export const ProductUpdate: React.FC<ProductUpdateProps> = ({ id, params }) => {
         closeModal();
         reset();
         refetch();
+        limitOpts.refetch();
       }
     }
   });
@@ -515,6 +522,7 @@ export const ProductUpdate: React.FC<ProductUpdateProps> = ({ id, params }) => {
         errors={errors}
         fetchCategories={searchCategories}
         fetchCollections={searchCollections}
+        limits={limitOpts.data?.shop.limits}
         saveButtonBarState={formTransitionState}
         media={data?.product?.media}
         header={product?.name}