+++ src/storybook/stories/products/ProductUpdatePage.tsx
@@ -2,7 +2,12 @@ import placeholderImage from "@assets/images/placeholder255x255.png";
 import { channelsList } from "@saleor/channels/fixtures";
 import { createChannelsData } from "@saleor/channels/utils";
 import { collections } from "@saleor/collections/fixtures";
-import { fetchMoreProps, listActionsProps } from "@saleor/fixtures";
+import {
+  fetchMoreProps,
+  limits,
+  limitsReached,
+  listActionsProps
+} from "@saleor/fixtures";
 import ProductUpdatePage, {
   ProductUpdatePageProps
 } from "@saleor/products/components/ProductUpdatePage";
@@ -47,6 +52,7 @@ const props: ProductUpdatePageProps = {
   hasChannelChanged: false,
   header: product.name,
   media: product.media,
+  limits,
   onAssignReferencesClick: () => undefined,
   onBack: () => undefined,
   onCloseDialog: () => undefined,
@@ -186,4 +192,10 @@ storiesOf("Views / Products / Product edit", module)
       )}
     />
   ))
-  .add("with channels", () => <ProductUpdatePage {...props} />);
+  .add("with channels", () => (
+    <ProductUpdatePage {...props} currentChannels={channels} />
+  ))
+  .add("no limits", () => <ProductUpdatePage {...props} limits={undefined} />)
+  .add("limits reached", () => (
+    <ProductUpdatePage {...props} limits={limitsReached} />
+  ));